Move ThumbsUtils and UriUtils to media/utils package
