+++ app/src/main/java/ir/mehdiyari/krypt/ui/media/utils/UriUtils.kt
@@ -1,4 +1,4 @@
-package ir.mehdiyari.krypt.utils
+package ir.mehdiyari.krypt.ui.media.utils
 
 import android.content.ContentResolver
 import android.content.ContentUris
